+++ app/models/public_file.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class PublicFile < ApplicationRecord
-
-  belongs_to :dataset
-  belongs_to :user
-
-end