+++ db/migrate/20160324131916_change_default_status_for_agreements.rb
@@ -1,0 +1,9 @@
+class ChangeDefaultStatusForAgreements < ActiveRecord::Migration
+  def up
+    change_column :agreements, :status, :string, null: false, default: 'started'
+  end
+
+  def down
+    change_column :agreements, :status, :string, null: true, default: nil
+  end
+end